# t1/smoke/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "T1 资料采集、清洗与元数据服务",
    "version": "1.0.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "summary": "Healthz",
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Healthz Healthz Get"
                }
              }
            }
          }
        }
      }
    },
    "/readyz": {
      "get": {
        "summary": "Readyz",
        "operationId": "readyz_readyz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Readyz Readyz Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/capabilities": {
      "get": {
        "summary": "Capabilities",
        "operationId": "capabilities_v1_capabilities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Capabilities V1 Capabilities Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sources": {
      "get": {
        "summary": "Sources",
        "operationId": "sources_v1_sources_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Sources V1 Sources Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/normalize": {
      "post": {
        "summary": "Normalize Documents",
        "operationId": "normalize_documents_v1_documents_normalize_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NormalizeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NormalizeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Chunk": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          },
          "sequence": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Sequence"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "text",
          "sequence"
        ],
        "title": "Chunk"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KnowledgeDocument": {
        "properties": {
          "source_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Source Id"
          },
          "source_uri": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Source Uri"
          },
          "media_type": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "audio",
              "video",
              "document"
            ],
            "title": "Media Type"
          },
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Title"
          },
          "authorization_status": {
            "type": "string",
            "enum": [
              "authorized",
              "public",
              "restricted",
              "unknown"
            ],
            "title": "Authorization Status"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "source_uri",
          "media_type",
          "title",
          "authorization_status"
        ],
        "title": "KnowledgeDocument"
      },
      "NormalizeRequest": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeDocument"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Documents"
          },
          "publish": {
            "type": "boolean",
            "title": "Publish",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "documents"
        ],
        "title": "NormalizeRequest"
      },
      "NormalizeResponse": {
        "properties": {
          "records": {
            "items": {
              "$ref": "#/components/schemas/NormalizedRecord"
            },
            "type": "array",
            "title": "Records"
          },
          "rejected": {
            "items": {
              "$ref": "#/components/schemas/RejectedDocument"
            },
            "type": "array",
            "title": "Rejected"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "records"
        ],
        "title": "NormalizeResponse"
      },
      "NormalizedRecord": {
        "properties": {
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "source_uri": {
            "type": "string",
            "title": "Source Uri"
          },
          "media_type": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "audio",
              "video",
              "document"
            ],
            "title": "Media Type"
          },
          "authorization_status": {
            "type": "string",
            "enum": [
              "authorized",
              "public",
              "restricted",
              "unknown"
            ],
            "title": "Authorization Status"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "chunks": {
            "items": {
              "$ref": "#/components/schemas/Chunk"
            },
            "type": "array",
            "title": "Chunks"
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "title",
          "source_uri",
          "media_type",
          "authorization_status"
        ],
        "title": "NormalizedRecord"
      },
      "RejectedDocument": {
        "properties": {
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "title",
          "code",
          "message"
        ],
        "title": "RejectedDocument"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}